moved docker swarm fixture to conftest
